refactor(conversations): extract action wrapper
Extracted ConversationListItemActionWrapper to a separate file for better
modularity and readability. Removed redundant imports and updated tests
accordingly.
